[Core] Sanitize default namespace for new files
Fixed bug # 59111 - New file added to project with hyphen in name has
invalid namespace
https://bugzilla.xamarin.com/show_bug.cgi?id=59111

Creating a new .NET Core project with a name that contains a hyphen
and then adding a new class file to the project would result in an
invalid namespace being used. The same thing could happen with a non
.NET Core project if the RootNamespace in the project file contains a
hyphen.