[Core] Re-evaluate AppliesTo condition on capability change (#2955)
* [Core] Add test for AppliesTo evaluation on capability change

This test checks that when a new project capability is added to a
project a project extension that uses an AppliesTo attribute is updated.

Also this test should not be in Core Tests since it uses the
AzureFunctions project extension.

* [Core] Re-evaluate AppliesTo condition on capability change

If the project extension uses an AppliesTo attribute which is
initially false and then a new project capability is added and the
project is re-evaluated then the AppliesTo attribute is now
re-evaluated. This will then cause the project extension to be added
to the extension chain. If the project extension was attached to the
project and the project capability is removed then the project
extension will now be removed when the project is re-evaluated.

* [Core] Re-create the extension context on refreshing the extensions

This avoids having to special case the AppliesTo attribute.