[Core] Fix MSBuild condition evaluation when using Czech locale
Fixed bug #58338 - New .NET Core project has net461 target framework
when primary Mac OS language is Czech
https://bugzilla.xamarin.com/show_bug.cgi?id=58338

Single.TryParse returns true for an empty string when the machine's
culture is 'cs-CZ'. This was causing the MSBuild condition evaluation
to be treated as a boolean comparison incorrectly in some cases.
The condition below would return false even though the property was
not defined and both sides of the condition were empty strings:

<PropertyGroup Condition="'$(TargetFrameworkIdentifier)' == ''">

This was causing the TargetFrameworkIdentifier to not be set for
.NET Core projects so the default of .NETFramework was being used.