Fix windows build (#3000)
* Output a binary log when building on Windows.

This is useful for investigating build issues.

* Fully qualify a type.

For some reason in Windows builds this fails to resolve because the namespace walk sees the "Microsoft.Runtime" namespace and fails. Not sure why it doesn't break on a Mac. Maybe that namespace isn't present on Mac inside the reference assemblies.

* Make sure MD.Core.Tests imports MonoDevelop.props.

Currently it has its own Signing properties and it doesn't specify PublicSign like all other projects. By doing what all other projects do we make sure this project is public-signed.

* Add a missing reference from WindowsPlatform.Tests to MonoDevelop.Core.Tests.