[NuGet] Detect official NuGet source using url not name
The System.Tuple NuGet package will now be available to install
as a quick action for the official NuGet package source even if
it has any source name. The url is now checked to see if the host
name ends with 'nuget.org' instead of checking for a source name
that exactly matches 'nuget.org'.