Fix Windows build. (#3070)
Don't resolve references from the output directory because there's no guarantee they will be there at the time the project builds. Instead resolve deterministically from the packages directory. It is annoying because Nuget forces you to specify the version in the HintPath, but hopefully with the new SDK projects and PackageReference this can go away soon.

Implement missing members on WPFToolbar.