[Core] Update solution startup item on re-evaluation
Fixed bug #58021 - New Azure Functions project can't be debugged
https://bugzilla.xamarin.com/show_bug.cgi?id=58021

When re-evaluating the project check to see if the project's
SupportsExecute has changed, which can happen if a new project
capability is installed via a NuGet package. If this has changed
then check to see if the solution's StartupItem should be refreshed
since it may now be possible to set a startup item for the solution.
Also handle the case where the project capability is removed and the
project can no longer be executed. In this case if the solution's
StartupItem is the project then the startup item should be changed.