[Core] Define MSBuildSDKsPath property for remote MSBuild engine
MSBuild on the command line defines the MSBuildSDKsPath globally in
its MSBuild.dll.config file. The MSBuild engine host that is used
when building in the IDE now defines also the MSBuildSDKsPath property.

This fixes a build error when using Xamarin.Forms 2.4.0 in an sdk style
project that targets .NET Standard. Xamarin.Forms 2.4.0 uses default
item imports which are not included since they are conditionally
imported:

<Import Project="$(MSBuildThisFileDirectory)Xamarin.Forms.DefaultItems.props" Condition="'$(MSBuildSDKsPath)'!=''" />

If the sdk style project has a .xaml file and a .xaml.cs file then
the xaml.g.cs file will not be generated when the project has no files
defined in the project since the default items are not imported. This
then causes a build error about the InitializeComponent method not
being defined.