[Core] Fix incorrect update metadata for wildcard item
Metadata defined for wildcard update items were being evaluated using
the wildcard item as the context instead of the expanded item. This
was causing .xaml.cs files to not be nested in the Solution window
for Xamarin.Forms 2.4.

Xamarin.Forms 2.4 defines an update item similar to:

<Compile Update="**\*.xaml.cs" DependentUpon="%(Filename)" />

The DependentUpon property was evaluated using the wildcard item
as the current context which resulted in the DependentUpon to be
evaluated as '*.xaml.cs' instead of the filename of the item that was
updated by the wildcard.