[Core] Fix evaluated update metadata from wildcard saved in project
Xamarin.Forms 2.4 defines a wildcard update similar to:

<Compile Update="**\*.xaml.cs" DependentUpon="%(Filename)" />

When a project was saved a Compile item was added to the main project
with the evaluated value stored in the DependentUpon element. Now
the evaluated value is treated as being the default value so it is
not saved in the main project.