[Core] Fix include wildcard metadata being saved in project file
With a wildcard include that defines MSBuild metadata on adding
a new file to the project an include MSBuild item with the metadata
would be saved in the project file. Also prevent a remove item being
added when the file is already removed by another glob.

An example is Xamarin.Forms 2.4 which defines wildcard includes:

<None Remove="**\*.xaml" />
<EmbeddedResource Include="**\*.xaml" SubType="Designer" Generator="MSBuild:UpdateDesignTimeXaml" />

Which would cause a new .xaml file added to the project and saved
in the project file as well as a remove for the None item even though
the .xaml file was already removed from the None items.

<None Remove="MyView.xaml" />

<EmbeddedResource Include="MyView.xaml">
  <Generator>MSBuild:UpdateDesignTimeXaml"</Generator>
</EmbeddedResource>