Switch some projects to use PackageReference instead of packages.config.
To make it easier to update VS Editor and Roslyn references in the future I'm switching to use <PackageReference> in the projects that use Roslyn and/or Editor.

One thing to know about PackageReference is that it works in non-SDK-style projects. A project can opt-in to supporting this either by having at least one PackageReference or setting <RestoreProjectStyle>PackageReference</RestoreProjectStyle>.

If ProjectB references ProjectA and both have opted in to use PackageReference, all PackageReferences from ProjectA will be available in ProjectB transitively.

In some projects you'll notice that I'm just adding a System.ValueTuple. This opts in and inherits the PackageReferences from MD.Core.

To make it easier to update versions I'm extracting the Editor and Roslyn versions into an MSBuild property and using that. This way we can bump the version by changing a single line.