[Core] Fix generated NuGet files being imported twice
The generated ProjectName.nuget.g.targets and ProjectName.nuget.g.props
that are created for .NET Core projects in the base intermediate
directory were being imported twice when evaluating the project.
Once by Microsoft.Common.props that is provided with Mono and once
with the imports that were added by the project model.

This fixes a duplicate file being added to the project's Files
collection when Xamarin.Forms 2.4 is used with a .NET Standard
project and no .NET Core SDK is installed. This would result in the
content page xaml and associated C# file not being nested in the
solution window.