[Core] Fix incorrect SDK style project updates when moving a file
Fixed bug #59418 - SDK project file not updated correctly when moving
file to different folder
https://bugzilla.xamarin.com/show_bug.cgi?id=59418

Moving a Resource file from one folder to another would not update
the project file correctly. Either the project file would not be
modified or the items would be removed from the project file.
One problem was that on moving a clone is taken of the original item
and the evaluated MSBuild item was copied but was still referring
back to the original location. Another problem was that MSBuild items
were incorrectly being removed when matching a null include.