[Core] Add support for multi-dimensional array writing.
The problem with the old code is that it assumed that any Array was uni-dimensional. Therefore, it would end up with an invalid cast exception.

There is a fast-path for rank1 arrays, as the foreach is rewritten to handle generalized iteration.

Bug 59805 - Visual Studio for Mac potential issue on unit testing support