[Core] Optimize SDK style object evaluation
The old code would create a list of objects to evaluate, resize the list by at least 2 elements and force a movement of all the elements in the list (which could be thousands) by 1 element, in a loop.
Use smarter Concat methods and force lazy evaluation, especially since GetAllObjects () is lazy.