[UnitTesting] Fix xUnit test messages not displayed
xUnit tests support an ITestOutputHelper interface which can be
passed to the constructor of the test. This has a WriteLine method
which can be used to output messages whilst running the test. When
this was used the text displayed in the Test Results window would
show:

Microsoft.VisualStudio.TestPlatform.ObjectModel.TestResultMessage

Instead of the actual text.