[Ide] Provide a way to get files from an item template
The ItemTemplate now has a GetStream method which takes a relative
path. This can be used to get the template.json file or a custom
host.json file.

  template.GetStream ("${TemplateConfigDirectory}/template.json");

The TemplateConfigDirectory placeholder can be used so the full
relative path to the template.json file does not need to be known for
each template.