[Core] Add file watcher for projects
File watcher can be enabled for a project by setting the
UseFileWatcher to true if the project has advanced glob support
enabled. Files created in the project directory or subdirectory
outside the IDE are added to the project and appear in the Solution
window. Deleting files and directories will remove the files from
the Solution window. Renaming files and directories is supported.
Files are added if they are included in file globs.

The Solution window does not currently show a directory after it
has been moved to a higher parent directory. For example, moving
'ProjectDir/Src/Test' to 'ProjectDir/Test'.