Fix incorrect completion popup position
When completion was triggered in an attribute value (="")
after typing the ", the completion window would attempt to
determine the position by calling TextViewMargin.ColumnToX,
which in turne would call TextViewMargin.CreateLinePartLayout.

Since the editor had just inserted the matching ", the line
would not be present in the cache. CreateLinePartLayout would
therefore build the line layout from scratch and then fire the
LineShown event.

LineShown appeared to exist purely to trigger
HighlightUrlExtension.Editor_LineShown. If there was an URL
in the line, this would add a marker... which immediately
invalidates the cache and destroys the layout.

Hence the layout created by TextViewMargin.CreateLinePartLayout
would become destroyed before being returned to its caller,
and throw exceptions when used.

However, the caller, TextViewMargin.ColumnToX SWALLOWS EXCEPTIONS
and simply returns 0. As a result, the popup location was wrong.

Fixed by switching LineShown to be fired _before_ the layout is
created and cached (and renaming it to LineShowing).

Also, log exceptions.