[build] fix issues bundling live player
Bump the size of the .dmg when creating it. There does not appear to be a suitable way to calculate
this accurately and playing trial and error trying to find a value that works is painful and likely
to break again. This is a hammer approach, but since all the extra space in the .dmg is compressed
away to nothing it doesn't really matter that much.