[Ide] Fix directory moved outside of IDE not updating Solution pad
When a directory in a .NET Core project was moved outside the IDE
to a parent directory inside the project (e.g. from ProjectDir/A/B to
ProjectDir/A) the directory would be removed from the Solution pad
and the new directory folder was not displayed. Moving a directory
to a child directory or renaming the directory would update the
Solution pad correctly. The problem was that the ProjectNodeBuilder
was using the FileService.FileRemoved to refresh the Solution pad.
Since the new directory had no associated ProjectFolder no tree
node could be found to refresh the UI. Now the FileService.FileCreated
event is used and the directory's parent is found and refreshed.