[DotNetCore] Fix file removed and re-added when file saved
Saving a C# file opened in the text editor would cause the file
to be removed and then added back to the project. The TextFileUtility
when writing a file will call FileService.SystemRename to move a
temporary file with the updated text to the original file's location.
This results in the file being deleted and then created. These
events were fired by the file watcher and resulted in the file
being removed and then added back to the project. Now a check is
made to see if the file does not exist before removing the file from
the project. The delete and create event happen close together so
this check, since it is made later on using the UI thread, seems to
happen after the original file has been created so it exists. This
prevents the file from being removed from the project.