Refactor code completion window
This commit refactors the code completion window and related infrastructure
to make it faster and easier to test. The new design is based in a MVC
pattern. The controller is implemented by the new CompletionController
class. The view is defined by the new ICompletionView interface, and
implemented by CompletionListWindowGtk.

All the logic in CompletionListWindowGtk, ListWindow and ListWidget classes
has been moved to a new CompletionController class. The ListWindow class
has been merged into CompletionListWindowGtk.

The performance of the completion window has been improved in different ways:

 * Removed unnecessary sort when initializing the completion window.
 * Match calculation results are now cached in completion items, and reused
   by the different operations done while showing the completion window.
 * Filtering can now be using several threads working in parallel when
   the list of items is very large.
 * Avoided unnecessary reallocations of the completion window.

New tests have been added and the testing framework is improved.

TimerCounter has some improvements to make it easier to trace time.

StringMatcher now has a clone method, required when using matchers
in multi-threading scenarios since matchers are not thread-safe.

Fixed bug in BacktrackingStringMatcher that caused the matcher to fail
when trying to match a string like 'aaa0' with pattern 'Aa'.