[Ide] Add IDE startup timer
Added a timer/counter for the IDE startup time. This is not using
a TimerCounter since it would be created/started before the any
instrumentation handlers are creates so the information would not
be available to these handlers. Instead a Counter is used and the
startup time is set as a property value in the counter's metadata.
Also recorded is whether a solution or document was opened on
starting the IDE.