[Ide] Fix empty folder not removed from solution window with watcher
VSTS 535727
Empty folder is not removed from solution explorer until we refresh
the solution/project

On removing a file externally if the folder was then empty the folder
was not removed from the Solution window when the folder was not
included in the project explicitly and the file watcher is enabled
for the project.

Also ensure the following scenarios work:

Uncollapsed folder and file removed - folder removed from solution
window.
Several levels deep folder containing single file which is deleted -
all empty folders removed from solution.
Folder explicitly included in the project file with an associated
MSBuild item is not removed when a file is deleted externally
from its folder and it is then empty.