Cleaned up the fix for dialogs
The problem with previous version of the fix is that it returns the dialog
result if it runs in sheet mode even if it's still running.

The workaround basically is to wait for the user's response to the dialog and
only then we can return results to client code.

To trigger the waiting loop I use RunModalForWindow method
and once the interaction is finished I destroy the loop with StopModal.