[Core] Indicate first build of project in instrumentation
Added a FirstBuild property to the project build timer. This is
true for a new project when it is built for the first time. The
FirstBuild value is stored in user prefs to handle the case where
the new project is created but not built and the solution is closed.