[Ide] Set open solution timer metadata in RootWorkspace
This avoids having to pass the metadata to the ProjectService's
ReadWorkspaceItem method. The metadata was originally passed to the
ProjectService so it was possible to indicate the solution was being
loaded at startup. Now the RootWorkspace creates and updates the open
workspace item timer metadata. Previously this was set on the
ReadWorkspaceItem timer.

One problem with using the OpenWorkspaceItem timer is when a solution
is contained inside a workspace. The ReadworkspaceItem timer generates
two timings, one for the solution and one for the workspace. The
OpenWorkspaceItem timer only generates a timing for the workspace.