[SourceEditor] Add typing counter
Counter records the time it takes to process a key press. The metadata
added to the counter is:

Extension - file extension
First - processing time for first key press in editor
Maximum - maximum processing time for all key presses in the editor
Average - average processing time for key presses.

The counter is updated when the text editor document is closed. All
recorded metadata is associated with the single document whilst it
is open. If there are no key presses then no counter information is
recorded.