[Core] Support .NET Core multi target framework using MSBuild property
VSTS 540666

An SDK style project that defined the <TargetFrameworks> element to
be another imported MSBuild property would fail to restore. The
unevaluated MSBuild property value would be used and result in an
unknown target framework being used causing the restore to fail.

<Project Sdk="Microsoft.NET.Sdk">
  <Import Project="Frameworks.props" />
  <PropertyGroup>
    <TargetFrameworks>$(TheFramework)</TargetFrameworks>
  </PropertyGroup>
</Project>

MSBuild property TheFramework is defined in the Frameworks.props file.

Now the evaluated property value is used for the TargetFrameworks
property.