[DotNetCore] Allow unknown target framework projects to be loaded
Fixed bug #57307 Unable to use Xamarin.iOS10 with TargetFramework
https://bugzilla.xamarin.com/show_bug.cgi?id=57307

The target framework of an SDK style project is no longer restricted.
This allows Tizen projects to be loaded without the error message
"Project does not support framework 'Tizen,Version=v4.0'" being
displayed. Tizen projects have a Tizen.NET package reference which
imports MSBuild .props file that defines the Tizen target framework.

The MSBuild.Sdk.Extras NuGet package is now also supported so
SDK style projects can use target frameworks that this NuGet package
supports, such as Xamarin.iOS and MonoAndroid.

On building an unknown target framework, that is not defined by
any used NuGet package references, a build error that indicates
the framework is not supported will occur:

The TargetFramework value 'test1.0' was not recognized.

This is the same behaviour as Visual Studio on Windows.

NuGet restore will also fail for an unknown target framework with
a message indicating the framework is invalid.