[Core] Fix project reference to shared project considered invalid
A project reference to a shared project that exists was shown as
invalid in the Solution window. It would have an error icon and have
the tooltip 'Project not found.' even though the shared project did
exist.

The problem was that the code was changed in commit d40ba5fec794edbe14ccf58c1e177944cd2a6fe9
so that if a project was not a DotNetProject then it was marked as
invalid even if it existed. The SharedAssetsProject class does
not derive from DotNetProject so the project was treated as though
it could not be found.