[Core] Fix issue with GetReferences working on an unsafe collection.
When calling `RunResolveAssemblyReferencesTarget` the returned list is apparently not guaranteed to be unmodified after it has been returned.

This causes issues for the rest of `GetReferences` because that list is iterated upon multiple times in the course of the method and thus is prone to receive InvalidOperationException because the collection has been modified during enumeration.

This commits ensure we always work on a copy of the data that's returned so that we can guarantee it won't be changed during execution.