[NUnit] Properly enumerate project assembly reference dependencies.
This commits changes the way dependencies of a test project are computed by relying on the MSBuild machinary instead of being limited to only the project model.

This fixes cases where assembly references are computed on the fly e.g. when custom `ReferencePath` are specified in the project so that `<Reference />` with no hint path are still resolved (something the old system would mistakenly think of as GAC assemblies).

The commits changes the API for gathering dependencies to be asynchronous and applies the same sort of filtering as before to only load dependencies that wouldn't be otherwise loaded.

An exception here is that facades are pulled in by the underlying call to `GetReferences` without a way to differentiate them from other references, that causes assembly load errors down the line which this commit works around that by wrapping the call in `InitSupportAssemblies` in a try/catch.

NB: The scenario that called for this fix is a MonoDevelop addin testsuite which has very specific considerations when it comes to assembly loading. Since that test-suite already uses a custom console runner (via vstool\mdtool) to execute unit-tests maybe it would be a good idea in general to also allow delegation of `GetTestInfo` to a custom runner as well.