Sync context fixes (#3624)
* [Core] Only set the main task scheduler once.

Possible fix for 533116

* [Ide] Modify DispatchService SynchronizationContext code so it does not allocate lambdas and delegates

* [Ide] Fix issue with the main GtkSynchronizationContext Send

Most of the time we were lucky, the Send was called by the GUI thread,
thus resulting in the fast-path where we already were on the GUI thread.

The problem was when this was being called from a background thread.

Monitor.Pulse can only be called by the lock holder. In this case,
nobody was holding the lock, just the bg thread was waiting on
the object which was owned by nobody.

I managed to reproduce this locally by forcing the Send to happen on a bg
thread and got this stacktrace:

```
System.Threading.SynchronizationLockException: Object is not synchronized
  at System.Threading.Monitor.ObjPulse (System.Object obj) [0x00008] in /Users/builder/jenkins/workspace/build-package-osx-mono-pullrequest/pr/external/bockbuild/builds/mono-x64/mcs/class/corlib/System.Threading/Monitor.cs:51
  at System.Threading.Monitor.Pulse (System.Object obj) [0x0000e] in /Users/builder/jenkins/workspace/build-package-osx-mono-pullrequest/pr/external/bockbuild/builds/mono-x64/mcs/class/referencesource/mscorlib/system/threading/monitor.cs:254
  at MonoDevelop.Ide.DispatchService+GtkSynchronizationContext+TimeoutProxy.HandlerInternal (System.IntPtr data) [0x00048] in /Users/therzok/Work/md/monodevelop/main/src/core/MonoDevelop.Ide/MonoDevelop.Ide/DispatchService.cs:82
  at (wrapper managed-to-native) Gtk.Application.gtk_main()
```

The other issue is that there's a small chance the code will race,
Monitor.Pulse _could_ happen before the Wait was done. Thus, you'd
end up with a thread being deadlocked.

Switching to a ManualResetEvent fixes this possible race.

* [Ide] GuiSyncContext now doesn't create overhead via tasks.

The old code would use the RunInMainThread code which generated wrapper tasks and ended up going only via the Post callback.
Use both SynchronizationContext methods, Post and Send to transmit data.