[Core] Fix UI hang on loading Xamarin.Forms solution
On loading the Xamarin.Forms solution, with NuGet restore enabled,
the UI thread would be stuck in a loop trying to refresh the project
folders. When a project uses a shared project the re-evaluation
would cause the shared files to be removed from the project. Then
an attempt was made to refresh a parent folder for that file being
removed, if the project had the file watcher enabled. Since the file
was outside the project directory the while loop would not finish
and cause the UI thread to be blocked.