Optimize project save (#3663)
Two optimizations: first, the default run configuration required for comparing
is generated only once and then cached. Generating the configuration is slow
because it needs a project evaluation.

Second, invalidation of the assembly and package references caches is now
instant. Instead of having to acquire a lock to clear the cache, there is now
a flag that can be set to signal that the cache is invalid and needs refresh.

Also added some logging.

This should fix VSTS issue #530382.