Convert MonoDevelop.AssemblyBrowser to use latest ICSharpCode.Decompiler.dll.
Takes the same version of ICSharpCode.Decompiler.dll (3.0.0.3403-beta4) that is used by the Roslyn we use.

Switch CSharpBinding to use the .dll.

Different way of getting a Cecil object for a type system object.

Use the AssemblyLoader instead of CecilLoader in places.

Call CSharpDecompiler.Decompile instead of AstBuilder.Add*

Rewrite how the actual text is generated (different visitors, etc)