[Ide] Fix broken automatic code generation
SourceTextContainer is not a TextBufferContainer for plain C# files. We currently don't have IVT access, so we cannot create TextBufferContainer in MD to be able to use it.
For now, use TryGetTextBuffer, as document change application in MonoDevelopWorkspace handles MonoDevelopHostDocumentRegistration.FromDocument returning null, falling back to old behaviour.
This fixes constructor generation