Switch to new ICSharpCode.Decompiler.dll
Switches for everything (Roslyn), but MonoDevelop.AssemblyBrowser.
AssemblyBrowser still references our own version of ICSharpCode.Decompiler.dll (0.0.0.0) and it is only copied to build/addins, not build/bin.

This is to ensure that both versions of Decompiler can co-exist side-by-side.

In the master branch a different approach is chosen: https://github.com/mono/monodevelop/pull/3682