[Core] Fix failure to create add new file to project
If a project has a default namespace of '1' then adding a new C#
file from a template to the project would throw a null reference
exception in the FileTemplateTagsModifier.ModifyTags method since
the namespace returned from DotNetProject.GetDefaultNamespace was
null. The problem was that the sanitized namespace for the default
namespace was null.

VSTS 524637