[NUnit] Fix load error displayed for ignored tests
When a test used an Ignore attribute the Unit Tests window would
show a load error for the entire project and prevent any tests from
being run. The problem was that the load error check was looking for
a _SKIPREASON property on every test-suite xml element in the xml
produced by NUnit v3. When an Ignore attribute is used the xml has a
_SKIPREASON property for the test case and that was being found and
treated as an error. To avoid this only a test-suite xml element that
has a type of Assembly and has a runstate of NotRunnable is checked
for a _SKIPREASON.

VSTS 544671