[Ide] Remove NuGet imports when migrating to project.json
When a PCL project was migrated to project.json the imports for
MSBuild .targets were not removed. These imports are added by NuGet
into the generated ProjectName.nuget.props and
ProjectName.nuget.targets files on restoring when a project.json file
is used. Having the import in the project file results in the import
being used twice. When a PCL project that uses Xamarin.Forms was
migrated to project.json the build would fail with an error:

Error XF001: Xamarin.Forms targets have been imported multiple times.
Please check your project file and remove the duplicate import(s).

Now on migrating the imports from the packages that were used by
the project are removed from the project file.

Fixes VSTS #552280