[Core] Fix Xamarin.Forms xaml file Remove item not removed
When an .xaml file was removed but not deleted, and Xamarin.Forms 2.5
used, which has default MSBuild items defined, on adding the .xaml
file back again to the project would result in the EmbeddedResource
Remove item not being removed from the project.

The problem was that the .xaml file was being added as a None item
since by default there is no build action specified for .xaml files.

Also another problem with Xamarin.Forms is that it defines Remove
items, that remove *.xaml files from the default None items that
the .NET Core sdk defines, in a separate .targets file from where
the original include items are defined. This resulted in the
Remove items being ignored since the file include globs were not
found.

Also fixed an msbuild include item added for a xaml file when it is
added from the file system. The Xamarin.Forms default msbuild item
for .xaml files have extra metadata which were not being added to
the file when it was added to the project from disk.