[Padtree] Disable the default treeview search feature
GtkTreeView has a feature that has an interactive search when you press
cmd/ctrl+f which we don't want to use. There is a bug in Gtk that reenables
the search when the model is set, so we need to hack around that

This fixes it for all uses of PadTreeView, including Search Results pad, Errors
pad, Breakpoints pad and many others.

Fixes VSTS #551327