[Core] Fix remove item not added for .xaml.cs file
With a .NET Core project that contains a single .xaml file and a
dependent .xaml.cs file, removing but not deleting the .xaml file
from the Solution window would not add an MSBuild remove item for
the .xaml.cs file even though it was removed from the Solution window.

The problem was that the Xamarin.Forms NuGet package includes a
Compile Update item and only this was being considered when saving
the project file. The default file glob, that includes all .cs files,
provided by the .NET Core SDK was not considered since it was not in
the unused items. Only the last MSBuild item associated with the
.xaml.cs file was included in the loaded items, and therefore the
unused items. If there was another .cs file in the project then the
Remove item was added correctly. To avoid this all MSBuild items
associated with a project item are added to the set of loaded items.