[Core] Allow project to stop items being excluded due to file glob
Added a UseDefaultMetadataForExcludedExpandedItems property to the
Project class. If this is true then when a new project item is added
and has missing properties that are included in an update glob then
an MSBuild include item will not be added to the project. If the
UseDefaultMetadataForExcludedExpandedItems is false, which is the
default, then the original behaviour occurs where an include item
is added. This feature will allow a Xamarin.Forms .xaml file to be
copied from a PCL project to a .NET Standard project without adding
an include item for the .xaml file to the project causing it to not
build due to a duplicate .xaml file.