[DotNetCore] Fix include item added for .xaml file copied from project
When copying a .xaml file from a PCL project to a .NET Standard project
an include item for the file would be added if the .xaml file did
not have the default metadata properties defined by Xamarin.Forms.
Now .NET Core projects opt-in to supporting items not being excluded
which prevents an MSBuild include item added for the .xaml file.