[Ide] Make root nodes retrieval dynamic
The addition of this code broke the live build output view (when building) as
when builders are started, the list of root nodes becomes invalid. So, make
it dynamic so that it always return the current status.

This is only called in one place (BuildOutputDataSource's constructor), so shouldn't
be any problem in enumerating the list of root nodes for each project whenever it is
called, which should happen when the build outputs have already been processed.