[Core] Fix xaml files not nested after editing project file in editor
Adding a new content page with xaml to a .NET Standard project,
then excluding the files from the project, but not deleting them,
then editing the project file and commenting out the MSBuild remove
items would then result in the xaml files not being nested in the
Solution window. The problem was that the MSBuild update item
for the .xaml.cs file, defined by the Xamarin.Forms default msbuild
items .targets, was being removed from its MSBuild project in
memory. This MSBuild update item has the DependsOn property defined
so this information was lost on reloading the project. Now a check
is to made to ensure only update items that exist in the user's
project are removed.

Fixes VSTS #505912