[Ide] Update dotnet templating engine version
Updated to version 1.0.0-beta3-20171117-314 which should be the
same as the one currently used in .NET Core SDK 2.1.4.

This fixes the following problems:

1) Templates that use files with '@' were being generated with the
@ symbol encoded.
2) Templates that use the Guid macro that do not specify a format
would cause the template generation to fail. An exception was thrown
since the format was not defined. With a later templating engine
if the format is not defined a default value is used.

Also changed how the template host is created by using the
AssemblyComponentCatalog. This simplifies the setup of the host and
also fixes the namespace not being updated in the generated
files. For example, in the .NET Core console project the Program.cs
would have a namespace of Company.ConsoleApplication1 instead of
using the project name.