[Ide] Support new lines in template descriptions
The template extension point used for the new templating engine uses
attributes for the description. To allow new lines the '\n' string
can be used and this will be replaced with a new line when the
description is displayed in the New Project dialog. If the description
needs to show '\n' but not as a new line then it can be escaped using
'\\n'.