[Ide] Don't run the completion service query on the UI thread (#3833)
* [Ide] Don't run the completion service query on the UI thread

The inner code in roslyn is thread safe, and we don't need to run the continuation on a bg thread either. Only sync back to UI context after the query is done

This amounts to 100ms out of 2000 of UI thread execution while moving the mouse

* [C#] LanguageItemTooltipProvider now does most of the CPU bound work on a bg thread

Out of 1.51s spent in the UI thread, 101ms were spent in this code, while moving the
mouse on screen.