[Core] Fix project capability test
The test was using the AzureFunctions project extension which has
been removed. Added a test addin that defines a custom project
extension using a test capability. This addin is registered by
adding an .addins file to the local-config directory. This file is
deleted after the test fixture completes to avoid including the addin
in the final packaged IDE.